# groupGifts-api-rest/src/main/resources/api.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /gift/{id}:
    get:
      tags:
        - gift-controller
      operationId: getGift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftDto'
    put:
      tags:
        - gift-controller
      operationId: updateGift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGiftDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftDto'
    delete:
      tags:
        - gift-controller
      operationId: removeGift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftDto'
    patch:
      tags:
        - gift-controller
      operationId: PartialUpdateGift
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartialUpdateGiftDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftDto'
  /gift:
    post:
      tags:
        - gift-controller
      operationId: createGift
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGiftDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GiftDto'
  /gift/filter:
    get:
      tags:
        - gift-controller
      operationId: findGifts
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: start_date
          in: query
          required: false
          schema:
            type: string
        - name: end_date
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GiftDto'


# CONTRIBUTION
...
components:
  schemas:
...
    GiftDto:
      required:
        - name
        - image
        - start_date
        - end_date
        - price
        - user_id
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        image:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        price:
          type: number
          format: float
        user_id:
          type: integer
          format: int64
    UpdateGiftDto:
      required:
        - name
        - image
        - start_date
        - end_date
        - price
        - user_id
      type: object
      properties:
        name:
          type: string
        image:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        price:
          type: number
          format: float
        user_id:
          type: integer
          format: int64
    CreateGiftDto:
      required:
        - name
        - image
        - start_date
        - end_date
        - price
        - user_id
      type: object
      properties:
        name:
          type: string
        image:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        price:
          type: number
          format: float
        user_id:
          type: integer
          format: int64
    PartialUpdateGiftDto:
      properties:
        name:
          type: string
        image:
          type: string
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        price:
          type: number
          format: float
        user_id:
          type: integer
          format: int64

# CONTRIBUTION
...